refactor(reports): move the report tables out of the design-code modules
Phase 3. The four _initialize_dicts_* builders and the _compile_results_* row
builders - 889 lines of labels, rounding and unit strings, about a third of each
code module - now live in mento/report_tables.py.

  codes/ACI_318_19_beam.py    1445 -> 938
  codes/EN_1992_2004_beam.py  1073 -> 612

The move was done by a script that cuts the exact source text rather than by
hand. 889 lines of UI strings is precisely where a transcription error hides,
and a numeric regression suite would not necessarily catch a wrong column
header.

More than relocation, the call moved too. _check_shear_* and _check_flexure_*
are calculation only now: they return None and no longer know that reports
exist. RectangularBeam._run_shear_check(force, report=...) decides whether to
build one, which is what lets shear_check_results skip the whole thing. The
report: bool flag that Phase 2b threaded through the code modules is gone with
it, and so is the EN shear row dict that was inline in its check.

report_tables.py is not in mypy's ignore_errors list, so 889 lines that were
previously unchecked are now type-checked. The six Optional errors that
surfaced were fixed - binding settings once through a narrowing helper - rather
than silenced by adding another ignore entry, which pyproject.toml explicitly
asks not to do.

The boundary tests now assert by AST that no beam code module defines a
_initialize_dicts_* or _compile_results_* function. Stale docstrings that still
promised a DataFrame return were corrected.

Remaining in this phase: matplotlib and docx are still inside beam.py, and the
wall report tables have not moved.

Suite 957 to 959 passed, coverage still 100%.
