+++ mento/codes/EN_1992_2004_beam.py
@@ -1,8 +1,6 @@
 import math
 from pint import Quantity
-from typing import TYPE_CHECKING, Dict, Any, Tuple, cast
-import pandas as pd
-from pandas import DataFrame
+from typing import TYPE_CHECKING, Tuple, cast
 
 # from devtools import debug
 
@@ -12,7 +10,7 @@ from mento.codes.en_1992_2004.equations import shear as shear_eq
 from mento.codes.flexure_design import _FaceDemand, _run_flexure_design
 from mento.material import Concrete_EN_1992_2004
 from mento.rebar import max_stirrup_spacing_EN_1992_2004
-from mento.units import MPa, mm, kNm, dimensionless, kN, N, inch
+from mento.units import MPa, mm, kNm, dimensionless, kN, N
 from mento.forces import Forces
 
 
@@ -185,7 +183,7 @@ def _calculate_required_shear_reinforcement_EN_1992_2004(
     self._V_Rd = self._V_Rd_s
 
 
-def _check_shear_EN_1992_2004(self: "RectangularBeam", force: Forces, *, report: bool = True) -> DataFrame | None:
+def _check_shear_EN_1992_2004(self: "RectangularBeam", force: Forces) -> None:
     if isinstance(self.concrete, Concrete_EN_1992_2004):
         if self._stirrup_n == 0:
             # Set current stirrup diameter to zero
@@ -233,31 +231,6 @@ def _check_shear_EN_1992_2004(self: "RectangularBeam", force: Forces, *, report:
             ) = max_stirrup_spacing_EN_1992_2004(self, self._alpha)
 
         self._DCRv = abs((self._V_Ed_2.to("kN").magnitude / self._V_Rd.to("kN").magnitude))
-
-        if not report:
-            return None
-
-        # Design results
-        results = {
-            "Label": self.label,  # Beam label
-            "Comb.": force.label,
-            "Av,min": round(self._A_v_min.to("cm ** 2 / m").magnitude, 2),  # Minimum shear reinforcement area
-            "Av,req": round(self._A_v_req.to("cm ** 2 / m").magnitude, 2),  # Required shear reinforcing area
-            "Av": round(self._A_v.to("cm ** 2 / m").magnitude, 2),  # Provided stirrup reinforcement per unit length
-            "NEd": self._N_Ed.to("kN").magnitude,
-            "VEd,1": self._V_Ed_1.to("kN").magnitude,  # Max Vu for the design at the support
-            "VEd,2": self._V_Ed_2.to("kN").magnitude,  # Max Vu for the design at d from the support
-            "VRd,c": round(self._V_Rd_c.to("kN").magnitude, 2),  # Concrete contribution to shear capacity
-            "VRd,s": round(self._V_Rd_s.to("kN").magnitude, 2),  # Reinforcement contribution to shear capacity
-            "VRd": round(self._V_Rd.to("kN").magnitude, 2),  # Total shear capacity
-            "VRd,max": round(self._V_Rd_max.to("kN").magnitude, 2),  # Maximum shear capacity
-            "VEd,1≤VRd,max": self._max_shear_ok,  # Check if applied shear is within max shear capacity
-            "VEd,2≤VRd": self._V_Ed_2.to("kN").magnitude
-            <= self._V_Rd.to("kN").magnitude,  # Check if applied shear is within total capacity
-            "DCR": round(self._DCRv, 3),
-        }
-        _initialize_dicts_EN_1992_2004_shear(self)
-        return pd.DataFrame([results], index=[0])
 
 
 def _design_shear_EN_1992_2004(self: "RectangularBeam", force: Forces) -> None:
@@ -579,7 +552,7 @@ def _design_flexure_EN_1992_2004(self: "RectangularBeam", max_M_y_bot: Quantity,
     _run_flexure_design(self, max_M_y_bot, max_M_y_top, _required, _capacity)
 
 
-def _check_flexure_EN_1992_2004(self: "RectangularBeam", force: Forces, *, report: bool = True) -> pd.DataFrame | None:
+def _check_flexure_EN_1992_2004(self: "RectangularBeam", force: Forces) -> None:
     """ """
     # Initialize the design variables requirements using the provided force.
     _initialize_variables_EN_1992_2004(self)
@@ -631,443 +604,9 @@ def _check_flexure_EN_1992_2004(self: "RectangularBeam", force: Forces, *, repor
     self._rho_l_bot = self._A_s_bot / (self._d_bot * self.width)
     self._rho_l_top = self._A_s_bot / (self._d_top * self.width)
 
-    if not report:
-        return None
-
-    # Compile the design results into a dictionary.
-    results = _compile_results_EN_1992_2004_flexure_metric(self, force)
-
-    # Initialize any additional dictionaries required for ACI 318-19 flexural checks.
-    _initialize_dicts_EN_1992_2004_flexure(self)
-
-    # Return the results as a Pandas DataFrame.
-    return pd.DataFrame([results], index=[0])
-
-
-def _compile_results_EN_1992_2004_flexure_metric(self: "RectangularBeam", force: Forces) -> Dict[str, Any]:
-    # Create dictionaries for bottom and top rows
-    if self._M_Ed >= 0:
-        result = {
-            "Label": self.label,
-            "Comb.": force.label,
-            "Position": "Bottom",
-            "As,min": round(self._A_s_min_bot.to("cm ** 2").magnitude, 2),
-            "As,req top": round(self._A_s_req_top.to("cm ** 2").magnitude, 2),
-            "As,req bot": round(self._A_s_req_bot.to("cm ** 2").magnitude, 2),
-            "As": round(self._A_s_bot.to("cm ** 2").magnitude, 2),
-            # 'c/d': self._c_d_bot,
-            "MEd": round(self._M_Ed_bot.to("kN*m").magnitude, 2),
-            "MRd": round(self._M_Rd_bot.to("kN*m").magnitude, 2),
-            "MEd≤MRd": self._M_Ed_bot <= self._M_Rd_bot,
-            "DCR": round(self._DCRb_bot, 3),
-        }
-    else:
-        result = {
-            "Label": self.label,
-            "Comb.": force.label,
-            "Position": "Top",
-            "As,min": round(self._A_s_min_top.to("cm ** 2").magnitude, 2),
-            "As,req top": round(self._A_s_req_top.to("cm ** 2").magnitude, 2),
-            "As,req bot": round(self._A_s_req_bot.to("cm ** 2").magnitude, 2),
-            "As": round(self._A_s_top.to("cm ** 2").magnitude, 2),
-            # 'c/d': self._c_d_top,
-            "MEd": round(self._M_Ed_top.to("kN*m").magnitude, 2),
-            "MRd": round(self._M_Rd_top.to("kN*m").magnitude, 2),
-            "MEd≤MRd": self._M_Ed_top <= self._M_Rd_top,
-            "DCR": round(self._DCRb_top, 3),
-        }
-    return result
+    return None
 
 
 ##########################################################
 # RESULTS
 ##########################################################
-
-
-def _initialize_dicts_EN_1992_2004_shear(self: "RectangularBeam") -> None:
-    if isinstance(self.concrete, Concrete_EN_1992_2004):
-        """Initialize the dictionaries used in check and design methods."""
-        self._materials_shear = {
-            "Materials": [
-                "Section Label",
-                "Concrete strength",
-                "Steel reinforcement yield strength",
-                "Safety factor for concrete",
-                "Safety factor for steel",
-                "Coefficient for long term effects and loading effects",
-            ],
-            "Variable": ["", "fck", "fywk", "γc", "γs", "αcc"],
-            "Value": [
-                self.label,
-                round(self.concrete.f_ck.to("MPa").magnitude, 2),
-                round(self.steel_bar.f_y.to("MPa").magnitude, 2),
-                self.concrete.gamma_c,
-                self.concrete._gamma_s,
-                self.concrete.alpha_cc,
-            ],
-            "Unit": ["", "MPa", "MPa", "", "", ""],
-        }
-        self._geometry_shear = {
-            "Geometry": [
-                "Section height",
-                "Section width",
-                "Clear cover",
-                "Longitudinal tension rebar",
-            ],
-            "Variable": ["h", "b", "cc", "As"],
-            "Value": [
-                self.height.to("cm").magnitude,
-                self.width.to("cm").magnitude,
-                self.c_c.to("cm").magnitude,
-                round(self._A_s_tension.to("cm**2").magnitude, 2),
-            ],
-            "Unit": ["cm", "cm", "cm", "cm²"],
-        }
-        self._forces_shear = {
-            "Design forces": [
-                "Axial, positive for compression",
-                "Shear",
-            ],
-            "Variable": ["NEd", "VEd,2"],
-            "Value": [
-                round(self._N_Ed.to("kN").magnitude, 2),
-                round(self._V_Ed_2.to("kN").magnitude, 2),
-            ],
-            "Unit": ["kN", "kN"],
-        }
-        # Min max lists
-        if self._V_Rd_s == 0 * kN:
-            self._stirrup_d_b = 0 * mm if self.concrete.unit_system == "metric" else 0 * inch
-        # Min max lists
-        min_values = [
-            None,
-            None,
-            self._A_v_min,
-        ]  # Use None for items without a minimum constraint
-        max_values = [
-            self._stirrup_s_max_l,
-            self._stirrup_s_max_w,
-            None,
-        ]  # Use None for items without a maximum constraint
-        current_values = [
-            self._stirrup_s_l,
-            self._stirrup_s_w,
-            self._A_v,
-        ]  # Current values to check
-
-        # Generate check marks based on the range conditions
-        checks = [
-            "✅" if (min_val is None or curr >= min_val) and (max_val is None or curr <= max_val) else "❌"
-            for curr, min_val, max_val in zip(current_values, min_values, max_values)
-        ]
-        self._all_shear_checks_passed = all(check == "✅" for check in checks)
-        self._data_min_max_shear = {
-            "Check": [
-                "Stirrup spacing along length",
-                "Stirrup spacing along width",
-                "Minimum shear reinforcement",
-            ],
-            "Unit": ["cm", "cm", "cm²/m"],
-            "Value": [
-                round(self._stirrup_s_l.to("cm").magnitude, 2),
-                round(self._stirrup_s_w.to("cm").magnitude, 2),
-                round(self._A_v.to("cm**2/m").magnitude, 2),
-            ],
-            "Min.": ["", "", round(self._A_v_min.to("cm**2/m").magnitude, 2)],
-            "Max.": [
-                round(self._stirrup_s_max_l.to("cm").magnitude, 2),
-                round(self._stirrup_s_max_w.to("cm").magnitude, 2),
-                "",
-            ],
-            "Ok?": checks,
-        }
-        self._shear_reinforcement = {
-            "Shear reinforcement strength": [
-                "Number of stirrups",
-                "Stirrup diameter",
-                "Stirrup spacing",
-                "Effective height",
-                "Minimum shear reinforcing",
-                "Required shear reinforcing",
-                "Defined shear reinforcing",
-                "Shear rebar strength",
-            ],
-            "Variable": ["ns", "db", "s", "d", "Asw,min", "Asw,req", "Asw", "VRd,s"],
-            "Value": [
-                self._stirrup_n,
-                self._stirrup_d_b.to("mm").magnitude,
-                self._stirrup_s_l.to("cm").magnitude,
-                round(self._d_shear.to("cm").magnitude, 2),
-                round(self._A_v_min.to("cm**2/m").magnitude, 2),
-                round(self._A_v_req.to("cm**2/m").magnitude, 2),
-                round(self._A_v.to("cm**2/m").magnitude, 2),
-                round(self._V_Rd_s.to("kN").magnitude, 2),
-            ],
-            "Unit": ["", "mm", "cm", "cm", "cm²/m", "cm²/m", "cm²/m", "kN"],
-        }
-        check_max = "✅" if self._max_shear_ok else "❌"
-        check_DCR = "✅" if self._DCRv < 1 else "❌"
-        rho_l = self._rho_l_bot if self._M_Ed >= 0 * kNm else self._rho_l_top
-        self._shear_concrete = {
-            "Shear strength": [
-                "Longitudinal reinforcement ratio",
-                "k value",
-                "Axial stress",
-                "Concrete strut angle",
-                "Concrete strength",
-                "Maximum shear strength",
-                "Total shear strength",
-                "Max shear check",
-                "Demand Capacity Ratio",
-            ],
-            "Variable": ["ρl", "k", "σcd", "Θ", "VRd,c", "VRd,max", "VRd", "", "DCR"],
-            "Value": [
-                round(rho_l.magnitude, 4),
-                round(self._k_value, 2),
-                round(self._sigma_cp.to("MPa").magnitude, 2),
-                round(math.degrees(self._theta), 1),
-                round(self._V_Rd_c.to("kN").magnitude, 2),
-                round(self._V_Rd_max.to("kN").magnitude, 2),
-                round(self._V_Rd.to("kN").magnitude, 2),
-                check_max,
-                round(self._DCRv, 3),
-            ],
-            "Unit": ["", "", "MPa", "deg", "kN", "kN", "kN", "", check_DCR],
-        }
-
-
-def _initialize_dicts_EN_1992_2004_flexure(self: "RectangularBeam") -> None:
-    if isinstance(self.concrete, Concrete_EN_1992_2004):
-        # Update longitudinal rebar attributes
-        self._update_longitudinal_rebar_attributes()
-        """Initialize the dictionaries used in check and design methods."""
-        self._materials_flexure = {
-            "Materials": [
-                "Section Label",
-                "Concrete strength",
-                "Steel reinforcement yield strength",
-            ],
-            "Variable": ["", "fck", "fyk"],
-            "Value": [
-                self.label,
-                round(self.concrete.f_ck.to("MPa").magnitude, 2),
-                round(self.steel_bar.f_y.to("MPa").magnitude, 2),
-            ],
-            "Unit": ["", "MPa", "MPa"],
-        }
-        self._geometry_flexure = {
-            "Geometry": [
-                "Section height",
-                "Section width",
-                "Clear cover",
-                "Mechanical top cover",
-                "Mechanical bottom cover",
-            ],
-            "Variable": ["h", "b", "cc", "cm,top", "cm,bot"],
-            "Value": [
-                self.height.to("cm").magnitude,
-                self.width.to("cm").magnitude,
-                self.c_c.to("cm").magnitude,
-                round(self._c_mec_top.to("cm").magnitude, 2),
-                round(self._c_mec_bot.to("cm").magnitude, 2),
-            ],
-            "Unit": ["cm", "cm", "cm", "cm", "cm"],
-        }
-        self._forces_flexure = {
-            "Design forces": [
-                "Top max moment",
-                "Bottom max moment",
-            ],
-            "Variable": ["MEd,top", "MEd,bot"],
-            "Value": [
-                round(self._M_Ed_top.to("kN*m").magnitude, 2),
-                round(self._M_Ed_bot.to("kN*m").magnitude, 2),
-            ],
-            "Unit": ["kNm", "kNm"],
-        }
-        # Min max lists
-        min_spacing_top: Quantity = max(
-            self.settings.clear_spacing,
-            self.settings.vibrator_size,
-            self._d_b_max_top,
-        )
-        min_spacing_bot: Quantity = max(self.settings.clear_spacing, self._d_b_max_bot)
-        min_values = [
-            self._A_s_min_top,
-            min_spacing_top,
-            self._A_s_min_bot,
-            min_spacing_bot,
-        ]  # Use None for items without a minimum constraint
-        max_values = [
-            self._A_s_max_top,
-            None,
-            self._A_s_max_bot,
-            None,
-        ]  # Use None for items without a maximum constraint
-        current_values = [
-            self._A_s_top,
-            self._available_s_top,
-            self._A_s_bot,
-            self._available_s_bot,
-        ]  # Current values to check
-
-        # Generate check marks based on the range conditions
-        checks = []
-        for i, (curr, min_val, max_val) in enumerate(zip(current_values, min_values, max_values)):
-            # --- EXCEPTION FOR DOUBLY REINFORCED SECTIONS ---
-            # If doubly reinforced, ignore maximum limits for top (i=0) and bottom (i=2)
-            if self._doubly_reinforced and i in (0, 2):
-                # If it passes min, we give the special tag
-                if min_val is None or curr >= min_val:
-                    checks.append("✅ D.R.")
-                    continue
-                # If it fails min, let the normal logic handle it (fall through)
-            # -------------------------------------------------
-
-            passed = (min_val is None or curr >= min_val) and (max_val is None or curr <= max_val)
-            if passed:
-                checks.append("✅")
-            else:
-                checks.append("❌")
-        self._all_flexure_checks_passed = not any(check in ("❌") for check in checks)
-        self._data_min_max_flexure = {
-            "Check": [
-                "Min/Max As rebar top",
-                "Minimum spacing top",
-                "Min/Max As rebar bottom",
-                "Minimum spacing bottom",
-            ],
-            "Unit": ["cm²", "mm", "cm²", "mm"],
-            "Value": [
-                round(self._A_s_top.to("cm**2").magnitude, 2),
-                round(self._available_s_top.to("mm").magnitude, 2),
-                round(self._A_s_bot.to("cm**2").magnitude, 2),
-                round(self._available_s_bot.to("mm").magnitude, 2),
-            ],
-            "Min.": [
-                round(self._A_s_min_top.to("cm**2").magnitude, 2),
-                round(min_spacing_top.to("mm").magnitude, 2),
-                round(self._A_s_min_bot.to("cm**2").magnitude, 2),
-                round(min_spacing_bot.to("mm").magnitude, 2),
-            ],
-            "Max.": [
-                round(self._A_s_max_top.to("cm**2").magnitude, 2),
-                "",
-                round(self._A_s_max_bot.to("cm**2").magnitude, 2),
-                "",
-            ],
-            "Ok?": checks,
-        }
-        check_DCR_top = "✅" if self._DCRb_top < 1 else "❌"
-        check_DCR_bot = "✅" if self._DCRb_bot < 1 else "❌"
-        self._flexure_capacity_top = {
-            "Top reinforcement check": [
-                "First layer bars",
-                "Second layer bars",
-                "Effective height",
-                "Depth of equivalent strength block ratio",
-                "Minimum rebar reinforcing",
-                "Required rebar reinforcing top",
-                "Required rebar reinforcing bottom",
-                "Defined rebar reinforcing top",
-                "Longitudinal reinforcement ratio",
-                "Total flexural strength",
-                "Demand Capacity Ratio",
-            ],
-            "Variable": [
-                "n1+n2",
-                "n3+n4",
-                "d",
-                "c/d",
-                "As,min",
-                "As,req top",
-                "As,req bot",
-                "As",
-                "ρl",
-                "MRd",
-                "DCR",
-            ],
-            "Value": [
-                self._format_longitudinal_rebar_string(self._n1_t, self._d_b1_t, self._n2_t, self._d_b2_t),
-                self._format_longitudinal_rebar_string(self._n3_t, self._d_b3_t, self._n4_t, self._d_b4_t),
-                round(self._d_top.to("cm").magnitude, 2),
-                self._c_d_top,
-                round(self._A_s_min_top.to("cm**2").magnitude, 2),
-                round(self._A_s_req_top.to("cm**2").magnitude, 2),
-                round(self._A_s_req_bot.to("cm**2").magnitude, 2),
-                round(self._A_s_top.to("cm**2").magnitude, 2),
-                round(self._rho_l_top.magnitude, 5),
-                round(self._M_Rd_top.to("kN*m").magnitude, 2),
-                round(self._DCRb_top, 2),
-            ],
-            "Unit": [
-                "",
-                "",
-                "cm",
-                "",
-                "cm²",
-                "cm²",
-                "cm²",
-                "cm²",
-                "",
-                "kNm",
-                check_DCR_top,
-            ],
-        }
-        self._flexure_capacity_bot = {
-            "Bottom reinforcement check": [
-                "First layer bars",
-                "Second layer bars",
-                "Effective height",
-                "Depth of equivalent strength block ratio",
-                "Minimum rebar reinforcing",
-                "Required rebar reinforcing top",
-                "Required rebar reinforcing bottom",
-                "Defined rebar reinforcing bottom",
-                "Longitudinal reinforcement ratio",
-                "Total flexural strength",
-                "Demand Capacity Ratio",
-            ],
-            "Variable": [
-                "n1+n2",
-                "n3+n4",
-                "d",
-                "c/d",
-                "As,min",
-                "As,req top",
-                "As,req bot",
-                "As",
-                "ρl",
-                "MRd",
-                "DCR",
-            ],
-            "Value": [
-                self._format_longitudinal_rebar_string(self._n1_b, self._d_b1_b, self._n2_b, self._d_b2_b),
-                self._format_longitudinal_rebar_string(self._n3_b, self._d_b3_b, self._n4_b, self._d_b4_b),
-                round(self._d_bot.to("cm").magnitude, 2),
-                self._c_d_bot,
-                round(self._A_s_min_bot.to("cm**2").magnitude, 2),
-                round(self._A_s_req_top.to("cm**2").magnitude, 2),
-                round(self._A_s_req_bot.to("cm**2").magnitude, 2),
-                round(self._A_s_bot.to("cm**2").magnitude, 2),
-                round(self._rho_l_bot.magnitude, 5),
-                round(self._M_Rd_bot.to("kN*m").magnitude, 2),
-                round(self._DCRb_bot, 2),
-            ],
-            "Unit": [
-                "",
-                "",
-                "cm",
-                "",
-                "cm²",
-                "cm²",
-                "cm²",
-                "cm²",
-                "",
-                "kNm",
-                check_DCR_bot,
-            ],
-        }
-        self._flexure_all_checks = (
-            self._all_flexure_checks_passed and (check_DCR_bot == "✅") and (check_DCR_top == "✅")
-        )
